Split admin and public fhir client api methods into separate controllers
